User Enemies, Node Changes
Made User Enemies which spawn when a user is created. They siege clean
nodes.
Made there be fewer, larger (in size, not cube quantity) nodes and
cubes, and changed the text to be around the node. Made the nodes rotate
too, rotating opposite directions for + or - change size.
Added arm and focuser onto wikiworld. Should change to smoothstep
movement eventually, but it works for now. Made nodes begin immediately
uploading as soon as the arm and focuser lock onto them.